Make UnoGame reusable
UnoGame will now always clear all of the piles on init() (always called
along with UnoGame#playGame, making it reusable. Some of the piles thus
had to no longer be final, resulting in a potential security
vulnerability, for which they have been made and should now be accessed
via their respective getter methods. Having to generate the draw pile
also meant that the UnoDeck must be stored in a separate field.