Fix a major bug in UnoStandardDeck
The bug in question had to do with the supposed cloning all UnoDeck
implementations have to do - UnoStandardDeck only cloned the list (by
creating a new ArrayList with the deck as a parameter), which in turn
created a shallow copy with shared fields. This is not expected and
causes unpredictable behavior.