Create UnoControlledGameTest and put stressTest into it
UnoControlledGameTest#stressTest is a simple yet pretty effective test
that plays 5000 rounds of UNO between two UnoHands. This in no way
guarantees that the game is bug-free and contains some random chance
(that is countered with the massive amount of rounds), but it in turn a
fairly broad test that indicates whether or not exceptions occur. It
also tests the stability of the card economy at the end of each round.