Fix major error in UnoProgressiveRulePack
Fixed a major error that caused an IllegalStateException if the amount
of cards to be drawn from was more than 1 and there weren't enough cards
in the draw pile. The issue had to do with the draw cards being closed
by the UnoDrawPile#mergeResetShuffle and then being closed again
(impossible, throws an IllegalStateException) by the subsequent
UnoDrawCard#drawTo operation. The fix ensures that all cards drawn are
closed first and then drawn to the hand (without the use of
UnoDrawCard#drawTo as that would force an attempt to close them again).